requirements.txt
# Python >= 3.10
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2
torch==2.1.2
sentence-transformers==2.2.2
transformers==4.36.2
faiss-cpu==1.7.2
gradio==4.12.0
tqdm==4.66.1
